Add files via upload
This data results from a preliminary simulation and is essentially a proof that the for-loop works. The simulation ran 100 replicates of 20 different sample sizes (1-20) and 30 different effect sizes (1-30) describing the difference between G and E. I'll use this data to build the code for the heat map, although it is likely that the parameterizations will end up changing.