Add files via upload
Data from first simulation. This simulation is essentially a proof of concept that the for-loop works. I ran 100 replicates for 20 different sample sizes (1-20) and 30 different effect sizes (1-30) denoting the difference between G and E. I'll use this data to build the heatmaps, but it is likely the parameterization will change in subsequent iterations.